+++ fe/fe-core/src/main/java/com/starrocks/analysis/Expr.java
@@ -65,12 +65,14 @@ import com.starrocks.thrift.TExpr;
 import com.starrocks.thrift.TExprNode;
 import com.starrocks.thrift.TExprOpcode;
 import com.starrocks.thrift.TFunction;
+
 import java.io.DataInput;
 import java.io.DataOutput;
 import java.io.IOException;
 import java.lang.reflect.Method;
 import java.util.ArrayList;
 import java.util.Collection;
+import java.util.Collections;
 import java.util.Iterator;
 import java.util.LinkedList;
 import java.util.List;
@@ -201,6 +203,8 @@ public abstract class Expr extends TreeNode<Expr> implements ParseNode, Cloneabl
 
     protected final NodePosition pos;
 
+    private List<String> hints = Collections.emptyList();
+
     protected Expr() {
         pos = NodePosition.ZERO;
         type = Type.INVALID;
@@ -236,6 +240,7 @@ public abstract class Expr extends TreeNode<Expr> implements ParseNode, Cloneabl
         fn = other.fn;
         printSqlInParens = other.printSqlInParens;
         children = Expr.cloneList(other.children);
+        hints = Lists.newArrayList(hints);
     }
 
     @SuppressWarnings("unchecked")
@@ -1175,7 +1180,7 @@ public abstract class Expr extends TreeNode<Expr> implements ParseNode, Cloneabl
                 // otherwise we may recurse infinitely.
                 Method m = root.getChild(0).getClass().getDeclaredMethod(NEGATE_FN);
                 return pushNegationToOperands(root.getChild(0).negate());
-            } catch (NoSuchMethodException|IllegalStateException e) {
+            } catch (NoSuchMethodException | IllegalStateException e) {
                 // The 'negate' function is not implemented. Break the recursion.
                 return root;
             }
@@ -1339,9 +1344,13 @@ public abstract class Expr extends TreeNode<Expr> implements ParseNode, Cloneabl
         this.fn = fn;
     }
 
-    public void setIgnoreNulls(boolean ignoreNulls) { this.ignoreNulls = ignoreNulls; }
+    public void setIgnoreNulls(boolean ignoreNulls) {
+        this.ignoreNulls = ignoreNulls;
+    }
 
-    public boolean getIgnoreNulls() { return ignoreNulls; }
+    public boolean getIgnoreNulls() {
+        return ignoreNulls;
+    }
 
     // only the first/last one can be lambda functions.
     public boolean hasLambdaFunction(Expr expression) {
@@ -1408,4 +1417,13 @@ public abstract class Expr extends TreeNode<Expr> implements ParseNode, Cloneabl
             return expr;
         }
     }
+
+    public void setHints(List<String> hints) {
+        this.hints = hints;
+    }
+
+    public List<String> getHints() {
+        return hints;
+    }
+
 }
