+++ fe/fe-core/src/main/java/com/starrocks/analysis/FunctionCallExpr.java
@@ -53,7 +53,6 @@ import com.starrocks.thrift.TExprNodeType;
 import java.io.DataInput;
 import java.io.DataOutput;
 import java.io.IOException;
-import java.util.Collections;
 import java.util.List;
 import java.util.Objects;
 import java.util.Set;
@@ -76,7 +75,6 @@ public class FunctionCallExpr extends Expr {
     // The slot SlotDescriptor nullable info will lost or change
     private boolean mergeAggFnHasNullableChild = true;
 
-    private List<String> hints = Collections.emptyList();
     // TODO(yan): add more known functions which are monotonic.
     private static final ImmutableSet<String> MONOTONIC_FUNCTION_SET =
             new ImmutableSet.Builder<String>().add(FunctionSet.YEAR).build();
@@ -377,7 +375,6 @@ public class FunctionCallExpr extends Expr {
         return true;
     }
 
-
     @Override
     public void write(DataOutput out) throws IOException {
         fnName.write(out);
@@ -485,11 +482,4 @@ public class FunctionCallExpr extends Expr {
         }
     }
 
-    public void setHints(List<String> hints) {
-        this.hints = hints;
-    }
-
-    public List<String> getHints() {
-        return hints;
-    }
 }
