+++ fe/fe-core/src/main/java/com/starrocks/sql/optimizer/operator/scalar/CallOperator.java
@@ -14,28 +14,20 @@
 
 package com.starrocks.sql.optimizer.operator.scalar;
 
-import com.google.common.base.Preconditions;
 import com.google.common.collect.Lists;
 import com.starrocks.analysis.FunctionCallExpr;
-import com.starrocks.analysis.FunctionName;
 import com.starrocks.catalog.AggregateFunction;
 import com.starrocks.catalog.Function;
 import com.starrocks.catalog.FunctionSet;
 import com.starrocks.catalog.Type;
-import com.starrocks.server.GlobalStateMgr;
-import com.starrocks.sql.analyzer.DecimalV3FunctionAnalyzer;
 import com.starrocks.sql.optimizer.base.ColumnRefSet;
 import com.starrocks.sql.optimizer.operator.OperatorType;
-import com.starrocks.sql.optimizer.rewrite.ScalarOperatorRewriter;
 
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
 import java.util.Objects;
 import java.util.stream.Collectors;
 
-import static com.starrocks.catalog.Function.CompareMode.IS_NONSTRICT_SUPERTYPE_OF;
-import static com.starrocks.sql.optimizer.rewrite.ScalarOperatorRewriter.DEFAULT_TYPE_CAST_RULE;
 import static java.util.Objects.requireNonNull;
 
 /**
@@ -59,7 +51,6 @@ public class CallOperator extends ScalarOperator {
     // Ignore nulls.
     private boolean ignoreNulls = false;
 
-    private List<String> hints = Collections.emptyList();
     public CallOperator(String fnName, Type returnType, List<ScalarOperator> arguments) {
         this(fnName, returnType, arguments, null);
     }
@@ -190,13 +181,6 @@ public class CallOperator extends ScalarOperator {
                 Objects.equals(arguments, other.arguments);
     }
 
-    public void setHints(List<String> hints) {
-        this.hints = hints;
-    }
-
-    public List<String> getHints() {
-        return hints;
-    }
 
     @Override
     public ScalarOperator clone() {
@@ -205,45 +189,11 @@ public class CallOperator extends ScalarOperator {
         List<ScalarOperator> newArguments = Lists.newArrayList();
         this.arguments.forEach(p -> newArguments.add(p.clone()));
         operator.arguments = newArguments;
-        operator.hints = Lists.newArrayList(hints);
         return operator;
     }
 
     @Override
     public <R, C> R accept(ScalarOperatorVisitor<R, C> visitor, C context) {
         return visitor.visitCall(this, context);
-    }
-
-    public static CallOperator buildMultiCountDistinct(CallOperator oldFunctionCall) {
-        Function searchDesc = new Function(new FunctionName(FunctionSet.MULTI_DISTINCT_COUNT),
-                oldFunctionCall.getFunction().getArgs(), Type.INVALID, false);
-        Function fn = GlobalStateMgr.getCurrentState().getFunction(searchDesc, IS_NONSTRICT_SUPERTYPE_OF);
-
-        ScalarOperatorRewriter scalarOpRewriter = new ScalarOperatorRewriter();
-        return (CallOperator) scalarOpRewriter.rewrite(
-                new CallOperator(FunctionSet.MULTI_DISTINCT_COUNT, fn.getReturnType(), oldFunctionCall.getChildren(),
-                        fn),
-                DEFAULT_TYPE_CAST_RULE);
-    }
-
-    public static CallOperator buildSum(ColumnRefOperator arg) {
-        Preconditions.checkArgument(arg.getType() == Type.BIGINT);
-        Function searchDesc = new Function(new FunctionName(FunctionSet.SUM),
-                new Type[] {arg.getType()}, arg.getType(), false);
-        Function fn = GlobalStateMgr.getCurrentState().getFunction(searchDesc, IS_NONSTRICT_SUPERTYPE_OF);
-        ScalarOperatorRewriter scalarOpRewriter = new ScalarOperatorRewriter();
-        return (CallOperator) scalarOpRewriter.rewrite(
-                new CallOperator(FunctionSet.SUM, fn.getReturnType(), Lists.newArrayList(arg), fn),
-                DEFAULT_TYPE_CAST_RULE);
-    }
-
-    public static CallOperator buildMultiSumDistinct(CallOperator oldFunctionCall) {
-        Function multiDistinctSum = DecimalV3FunctionAnalyzer.convertSumToMultiDistinctSum(
-                oldFunctionCall.getFunction(), oldFunctionCall.getChild(0).getType());
-        ScalarOperatorRewriter scalarOpRewriter = new ScalarOperatorRewriter();
-        return (CallOperator) scalarOpRewriter.rewrite(
-                new CallOperator(
-                        FunctionSet.MULTI_DISTINCT_SUM, multiDistinctSum.getReturnType(),
-                        oldFunctionCall.getChildren(), multiDistinctSum), DEFAULT_TYPE_CAST_RULE);
     }
 }
