+++ fe/fe-core/src/test/java/com/starrocks/analysis/SelectStmtTest.java
@@ -273,12 +273,15 @@ public class SelectStmtTest {
     }
 
     @Test
-    public void testGroupByAllConstants() throws Exception {
+    public void testGroupByCountDistinctWithSkewHint() throws Exception {
         FeConstants.runningUnitTest = true;
-        starRocksAssert.getCtx().getSessionVariable().setOptimizerExecuteTimeout(1000000000);
-        String sql = "select cast(k1 as int), count(distinct [skew] cast(k2 as int)) from db1.tbl1 group by cast(k1 as int)";
+        String sql =
+                "select cast(k1 as int), count(distinct [skew] cast(k2 as int)) from db1.tbl1 group by cast(k1 as int)";
         String s = starRocksAssert.query(sql).explainQuery();
-        System.out.println(s);
+        Assert.assertTrue(s, s.contains("  3:Project\n" +
+                "  |  <slot 5> : 5: cast\n" +
+                "  |  <slot 6> : 6: cast\n" +
+                "  |  <slot 8> : CAST(murmur_hash3_32(CAST(6: cast AS VARCHAR)) % 512 AS SMALLINT)"));
         FeConstants.runningUnitTest = false;
     }
 }
